separate elsif branch for one SQL/ISO format
This format we use when fetch from DB. All fields in the string
are mandatory and it's treated as UTC time.

For this particular and very common case we can do
faster conversion to unix time. Other elsif branch
still handles more variants.

Our existing tests dictate that we normalize old dates
(negative unix time) to zero, as well as, invalid dates.
I think that we can avoid validation and use timegm_nocheck.