Convenience method to load arbitrary role groups based on an object
Rather than requiring the caller to discern the class of object and call
one of the class specific load methods, just pass in the object itself
and be happy.

RT::Test->add_rights now uses this and as such supports granting rights
to role groups on objects other than queues or the global system.