Display date format options in the user's timezone
It is now obvious from the formats which ones are defined as always
being in UTC vs. the ones which happened to be displayed that way in
preferences.  See [rt3 #20856] for an example of confusion caused by all
UTC datetimes.

Despite no user visible change, switch from ->Set to ->SetToNow because
the effect is much more clear when reading the code.