Test plugin loading and lib overlays in plugins via rt-server
Currently failing because RT::User is loaded too early.  Requires
testing under apache because otherwise the RT init sequence in the test
process (via RT::Test->import) interferes with rt-server in a way that
causes it to pass.

It is far too easy to break certain lib overlays when provided via
plugins, so hopefully a simple test will help.