Rewrite RT_SiteConfig.pm entirely when setting options during testing
Because only the first Set(...) takes effect, tests which set a value
twice failed when run under Apache -- the only time that the external
RT_Siteconfig is used during testing.  Instead of blindly appending to
the file, read it back in in its entirety, strip out any previous
Set(...) calls to the same variable, and then re-write it.