Move locale loading earlier, so we don't cache false negative
On the first call to Locale::Maketext->get_handle, the list of locales
is inspected, and RT::I18N::xx is attempted to be loaded to each
possible handle -- and load success/failure is cached indefinitely.
Thus, if a localization happens before RT::I18N->Init has been called
(which creates those classes), RT will forever assume that it does not
have any of the tried localizations.

This is not an issue during real deployment, as RT::I18N->Init is called
in RT->Init, which guaranteedly comes before any localizations that RT
might try.  However, in tests, we bootstrap the database into existance
in RT::Test->import _before_ calling RT::I18N->Init, which (in forked or
in-process test servers) caches a false negative for the 'en'
localization.  This causes "Homepage" to not be localized to "Home" in
the default locale.

Note that non-en localizations are unaffected, as they are not tried
during the initial bootstrap process, and hence they do not get a false
negative cache.

During testing, load the I18N files just before bootstrapping the
database.