Cleaner POD → HTML generation using more subclasses
Subclassing not just Pod::Simple::HTMLBatch but also Pod::Simple::XHTML
and Pod::Simple::Seach gives us more fine-grain options in overriding
discovery and output.  The end result is less copying of the internals
of HTMLBatch, which will be more maintainable.

Our HTMLBatch subclass writes its own table of contents (index.html) for
the generated doc so that we can exercise full control without dealing
with the absurd implementation details of HTMLBatch's contents
generation.

We now look in local/ directories for POD as well, which will help for
plugins.