Shredder: Loop until all objects on the stack are WIPED
Visit every object pushed onto the stack and wipe it.  This causes test
failures because resolvable dependencies (with flags DEPENDS_ON |
VARIABLE) are pushed onto the stack as well and should not be wiped.

The previous fix, 5bdf351, switched from "while each" to "for values" to
avoid the well-known side-effects of using each() when the hash is
modified inside the loop (pushing on new dependent objects).  That
commit fixed "skipped object" problems caused by an unpredictable
each(), but didn't address the problem of objects pushed onto the stack
after "values" was evaluated.