Shredder: Don't push resolvable dependencies onto the stack for wiping
If a dependency is marked VARIABLE (resolvable), require it is handled
by a resolver.  The resolver is free to put the object on the stack
explicitly if resolution is not actually possible.

Such objects were often safe because of the accidental side-effects of
modifying the hash iterated by each() and later only calling values()
once.  It is cleaner and safer to simply never push them onto the stack
in the first place.  This lets you re-use shredder objects, for example,
without inadvertently wiping objects you didn't mean to on the second
call to WipeoutAll.