cache_self doesn't play nice with our use of flush_buffer; remove it
Calling flush_buffer explicitly (as RT does in many places) causes
cache_self to only cache partial content.  Specifically, it only caches
the content generated since the last call to flush_buffer.  This is a
Mason bug[1] which, despite a patch and tests, will not be fixed by the
authors due to Mason 1's status as maint-only and the complexity of
Mason's buffering and caching system.

An alternate solution removes all (or a few strategic) calls to
flush_buffer, but that's a more widely affecting change.  This GnuPG
admin element shouldn't be used terribly much on a day to day basis, and
being a little slower because of the removal of cache_self is an
acceptable trade-off.

[1] Mason bug: https://rt.cpan.org/Ticket/Display.html?id=78754