Make File::Temp UNLINK setting explicit
File::Temp has a default UNLINK setting of 1, which means it
will delete temp files when the object goes out of scope. This
change explicitly sets the UNLINK value to false when in
debug mode, and leaves it true otherwise.

File::Temp processes the unlink operation in its DESTROY method
which means that in practice, the unlink code is typically not called
when persistent processes like the RT standalone server or an
apache process are killed and DESTROY is not called.
However, for cases where the RT code exits normally or dies, email
created with ($MailCommand, 'testfile') can be missing unexpectedly.
Running rt-crontool, for example, with an action that sends mail shows
this behavior.