run scrip which condition isn't UserDefined only on tickets of type ticket
With this change the predefined conditions in the select list are only
applicable on tickets of type ticket as the condition description notes.

For example:
On an vanilla RT installation if you create an reminder the queue AdminCcs are
notified because of the 'On Create Notify AdminCcs' scrip which isn't always
the desired behavior.

If you want to run an scrip on other than tickets you have to use
the UserDefined condition and can control in custom condition on
which ticket types this scrip should run.