deprecate /Ticket/Elements/EditCustomFields
In favor of /Elements/EditCustomFields.

Call on create is slightly longer. Note that passing
new custom fields collection is done to avoid re-applying
limits on the same collection.

It's deperecated not only because of its simplicity, but
also because of new /Elements/EditCustomFieldCustomGroups
comp which calls /Elements/EditCustomFields directly skiping
component in /Ticket/Elements. So any customization the subject
wouldn't be applied to all CF editors.