Check the original request for side-effects before prompting for login
Sets the foundation for informing users about to login what will happen
after they do so in order to help prevent phishing.

Note that while this shares some code and config with CSRF, it _isn't_
CSRF since the user is not yet logged in and it requires a user to fill
in valid credentials.  Sharing config allows administrators to whitelist
external sites, such as corporate portals, which deep-link into RT, thus
preventing RT from crying wolf to the end user about legitimate
requests.