Inform the user logging in about potential side-effects
This helps prevent phishing where the user follows a malicious link and
isn't logged in yet.  Previously there was no indication of what would
happen after login.  The CSRF protection does somewhat double duty and
provides the same measure of phishing protection when the user is
already logged in, but it is never triggered by unauthenticated
requests.

The warnings on login are controlled by the same configuration as CSRF.