Let MIME::Head modify the X-RT-GnuPg-Status header to handle continuations
X-RT-GnuPG-Status is a set to a many-line string and previously
continuations weren't handled properly.

Temporarily letting the inserted headers be modified fixes the known
problem.  If instead we marked the entire entity as modifiable early
after parsing and left it that way, the possibility of unforeseen
breakage would be too high for a stable release series.  It would also
lead to even further deviation of RT's stored message from the original
source message.