Make fake web request methods safer
The fake methods were returning the object instance itself, rather than
a data type compatible with what the real object does.  While this does
not cause any bugs at this point, it could lead to unexpected
consequences in the future; for example anything calling headers_out in
an array context would break, even though it's a valid thing to do.