Remove the IE-specific need for two <table> attributes on collection lists
The only reason IE padding was wonky without the cellpadding attribute
was because our stylesheet set up different padding for IE.  The blanket
padding is seemingly unnecessary.

The border="0" attribute is unnecessary since a border is set in CSS.

Only cellspacing is out of the reach of CSS in < IE8 because
border-spacing isn't supported.

Less hardcoded attributes in HTML means more flexibility for CSS down the
road.  While this is effectively a no-op commit in terms of rendering,
it's useful to avoid reinstating the two extra <table> attributes.