Remove unnecessary RT::Class::_LookupId
It was previously required because the class was RT::FM::Foo, and hence
the generic method in RT::Record did not work (as noted in the comment
there).  With the merge of RTFM into core, and the rename of
RT::FM::Class -> RT::Class, this method is no longer necessary.