Replace form_name/input/click pattern with a simple usage of submit_form_ok
The "with_fields" form selection is more robust than relying on the form
name, which may change when this branch is merged.  Additionally,
submit_form_ok validates that it was able to a) find a form, b) find the
fields to fill in, and c) click the button.  None of previously used
methods were error checked.

Two uses of form_name still exist, but they are a slightly more complex
pattern of selecting a form and building up inputs in a loop.
Additionally, the DOM is checked more rigorously for correct inputs in
those test blocks.

For ease of development, switch to a "done_testing()" approach to the
test plan rather than a static count.