Rename --devel-mode to --developer, to make break in meaning clear
Previously, --devel-mode both added the dependencies needed for testing
RT, as well as setting the default value of $DevelMode in RT_Config.pm
These actions are orthogonal, as many users wish to enable $DevelMode
but do not need the additional dependencies; and developers may not wish
to run with $DevelMode.

By renaming the option, existing users of --devel-mode will no longer
have the additional dependencies if all they wish is Module::Refresh.
This also makes the setting of $DevelMode in RT_Config.pm no longer
controlled at configure-time, which is an odd action-at-a-distance, as
it means one RT's defaults are malleable from one install to another.
Consistency is improved by forcing users to add it to RT_SiteConfig.pm,
like any other option.