ShowLinks: Construct clone links only if the current user can CreateTicket
Instead of separating the CreateTicket check from the URL construction,
simply skip constructing any clone URLs if the user can't CreateTicket
and conditionalize the display of clone links on the presence of the
constructed URLs themselves.

A cleaner separation of concerns is useful in preparation for ShowLinks
working on other objects other than Tickets.