Bulk update: Only show statuses for queues which appear in the results
Trims down the status selector from "all statuses" to a potentially much
smaller set in many scenarios.  It's safe to limit selectable statuses
because Status is set _before_ Queue when processing the update, so all
valid statuses will be listed.  If Queue is ever processed before
Status, this change will need to be revisited as it prevents showing
statuses from destination queues which don't appear in the search
results.