Group by lifecycle in the status selection widget
For UI simplicity, no grouping is done if only one lifecycle is seen.

For code simplicity, multiple <option> elements may be marked selected
even if the displayed <select> doesn't allow multiple selections.  This
is an authoring error according to the HTML spec, but browsers generally
just deal and pick one.  It doesn't matter to RT which one they pick,
since statuses are identified only by name.  We can consider a technical
solution down the road if it proves to trigger confusing behaviour and
frustrates users.