Simplify status selector by lumping lifecycles with identical statuses
For example, the "default" lifecycle and a copy named "reply-on-resolve"
which changes only the transitions available in the actions menu are now
grouped into the same <optgroup> with the heading "default,
reply-on-resolve".

While the previous commits made the shipped "approvals" and "default"
lifecycles appear separately if you were looking at tickets in both
queues, this commit now joins them.  For maximum simplicity, the
decision to display groupings or not is done _after_ simplification,
regardless of how many lifecycles may be lumped into the same grouping.