Ignore symlinks in the testing tree when using get_relocatable_dir/file()
Parent directory traversals (C<..> or File::Spec->updir()) are naively
canonicalized based on the test file path (C<$0>) so that symlinks
aren't followed.  This is the exact opposite behaviour of most
filesystems and is considered "wrong", however it is necessary for some
subsets of tests which are symlinked into the testing tree.