Record a txn on the RT::Record object when adding or removing a role member
A transaction is already recorded on the RT::Group itself, but it's
useful to have a transaction directly on the RT::Record too.

Doing this in the generic methods is easier than subclass-by-subclass
since AddRoleMember already validates PrincipalId and translates from
acceptable User/Group arguments.  Subclasses which don't want to record
a transaction can simply wrap the methods to force Silent => 1.