Add cflabel to mobile style sheet
Add cflabel to mobile label styles so custom fields on the mobile
reply page look the same as core labels. Although cfs aren't
normally updated on mobile, this was needed for
RT::Extension::MandatoryOnTransition and will likely be needed for
any other changes that need to display an editable cf on a mobile page.