Remove support for intermixing transactions from multiple tickets in history
Not supporting this anymore simplifies a lot of the ticket attachment
and transaction finding code.  Core RT never used the functionality; it
was added back in 2004 by 15fb328 to support an extension for a single
customer.  Many places displaying history also neglected to support it
since the functionality was in Ticket/Display.html instead of an
Element.  Any future extensions which need to accomplish the same result
may overlay/extend the FindTransactions element and the Attachments
method in RT::Ticket instead of using callbacks to set a collection
object.

Note that this wasn't used for merged tickets; those are handled
simply by limiting on EffectiveId instead of id.

/Ticket/Elements/FindAttachments is now deprecated.