Push the CF grouping into the NamePrefix, to allow for separation
If a custom field is rendered in multiple groupings on one page, and is
a select-multiple CF, the values between the two will be intermingled;
if a single-valued CF, the rendered defaults will become ARRAY(0xBEEF).

Avoid this by splitting -Value, -Magic, etc, on a per-grouping basis.
This allows us to associate them on a per-widget basis on the back-end,
and at least ensure that a consistent value is seen from what is
submitted.  In this case, we arbitrarily choose the alphabetically first
grouping as the one whose value is used.