Increase Log::Dispatch minimum version to 2.30, for consistent warnings
Prior to Log::Dispatch version 2.30, the "warn" alias for the "warning"
method failed to correctly canonicalize the level internally, and thus
failed to ever usefully generate output.  All of the warnings fixed in
224b96e thus only appear on Log::Dispatch 2.30 and above, due to this
bug -- and consequently the newly-added 'warning_like' tests fail with
earlier versions.

Bump the minimum required Log::Dispatch version to one which
consistently generates warnings with both ->warn and ->warning.