Escape '/'s in JSON, to allow safe use in <script> tags
The JSON produced by the JSON module is unsuitable for insertion into
HTML; user-supplied data could contain the string '</script>' which
would immediately close the surrounding <script> tag and allow for XSS
injection.

Escape all '/'s into '\/', a transformation which is allowed by the JSON
syntax.  This prevents insertion of </script>, by transforming it into
the HTML-harmless token <\/script>.  While JSON::PP has an option for
this, ("escape_slash") it is not currently supported by JSON::XS; as
such, implement escaping of '/'s by way of a regular expression.