Stop splitting SetOwner into _Set and _NewTransaction
This split, along with the comment that goes with it, originated in
4137c44, where the ->Commit statement occurred between two calls to
->_Set, one which changed the record, the other which added a
Transaction.  This was due to scrips, at the time, needing to be
triggered outside of a database transaction, due to the lack of nested
transactions.

Long after this restriction had been lifted, 6061508 moved ->Commit to
the more sensible place at the end of the updates, but left the two
calls to _Set (one of which had changed into _NewTransaction in e04e4a0)
in place, along with the misleading comment.

Replace both calls with one straightforward call to ->_Set.