Remove UpdateTicket argument to _Set, simplifying logic
The last existing call site which used UpdateTicket => 0 was removed in
e04e4a0 -- which was the call that UpdateTicket's functionality was
originally added for in 4137c44.

Remove the argument, and simplify the logic of _Set considerably.  The
equivalent functionality can be accomplished via _NewTransaction, which
is what other codepaths do when similar effects are desired.