Update denormalized "Owner" column on tickets on group change
Rather than explicitly handle the Owner column in all places which alter
the Owner role group, generalize that update into the group.  It is thus
sufficient to update the role group membership, and the denormalized
column will be updated appropriately.

Though the object in question can be determined via the properties of
the role group, we pass the object currently being examined such that
the data structure in memory is updated with the new value -- otherwise,
it stores stale data, leading to inconsistent and conflicting results.

It is important to note that the ticket object is created with an empty
Owner field, which is set by the ->_AddMember call later.  Failure to do
so causes inconsistency between the group membership and the field.