Make CreateRoleGroup respect InsideTransaction
Contrary to its documentation, CreateRoleGroup did not respect the value
of InsideTransaction that was passed to it, forcing it to be enabled.
While the only call site which disabled InsideTransaction was during
database bootstrapping (during which you have bigger problems if role
group creation fails), update the code to be consistent with the
documentation.