Factor out role group argument parsing
The code to parse Owners and Cc/AdminCc/Requestors is more generally
applicable than to tickets, and holds for any single-user role and other
role groups.  Abstract this code into a method on RT::Record.

This temporarily removes the deferred ownership feature, as well as
(relatedly) hiding error messages caused by the proposed owner not
having OwnTicket.  These shortcomings are addressed in the following
commit.