Pass Default value to SelectDate
The Default value is passed to EditCustomFieldDate and
EditCustomFieldDateTime, but they don't pass it along to
SelectDate. Without it, any value provided is lost when reloading
the page but not submitting, like after adding an attachment.

In SelectDate, Value was being defaulted with ||, but this
incorrectly set it to undef when Value was the empty string
and Default was undef.