Stay on the same rights subtab when switching between principals
By hooking the "show" event instead of the "select" event we avoid an
infinite loop when we then select the new tab in each tabset: by the
time "show" fires, the selected tab is already correct.

Based on an initial implementation by sunnavy, which differed primarily
in hooking the "select" event and unregistering/re-registering to work
around the loop.

Resolves [rt3 #19769].