Preserve the selected rights category tab across page reloads
For example: granting a right to "Privileged" from the "Rights for
Administrators" tab will return you to both tabs after submission.

We still need the anchor framework for the principal tabs in order to
translate newly added principals to their correct tab server-side.  The
non-newly-added-principal side could be replaced with similar cookie
usage, but it seems not as worthwhile since we'll still be stuck
wrangling anchors.  One slightly verbose method is better than two
competing methods.