Provide transaction brief descriptions as HTML
None of the descriptions are actually HTML yet, but the method is
available and waiting for descriptions to be HTML-ified.

Plain text brief descriptions are now scrubbed versions of the HTML for
consistency and single point of maintenance.

Instead of calling loc() themselves, the description handlers now return
a list which is passed to loc() after being properly escaped.  The
handling is similar to column map return values: scalar refs are passed
through unescaped, array refs are recursively processed and joined into
a single value, and everything else is HTML escaped.  This makes it easy
to start providing HTML descriptions without the hassle of escaping and
concatenation everywhere.

One negative side-effect of this is that all the localizable strings
will lose their example parameters when extracted into the message
catalog.  We can remedy this in the future by providing a comment mark
like #loc that also captures following parameters.