Add an AJAX endpoint for requesting user information and formatting
Javascript which may want to display a formatted user now has a way to
fetch the preferred format, as well as retrieve user details.  Multiple
users may be requested in a single call by specifying multiple values
for the "id" parameter.  A JSON hash is returned with keys matching the
requested ids for easy client side lookup.  The JSON object looks like:

    {
       "12" : {
          "id" : "12",
          "Name" : "root",
          "RealName" : "Enoch Root",
          "EmailAddress" : "root@localhost",
          "Privileged" : true,
          "_formatted" : "Enoch Root"
          "_html" : "Enoch Root",
       }
    }