Make ValidateWebConfig account for proxied hosts
Many reverse proxies alter the headers that they pass through to the
backend server, which causes spurious warnings from ValidateWebConfig.
For example, while WebDomain is set up to point to the name of the
reverse proxy host, the Host: header that is received (from the proxy)
is the name of the back-end server.

Look at the pre-proxied header values, when they exist, to determine if
WebPort and WebDomain are configured correctly.  Also consolidate the
two separate WebDomain checks, for clarity.