Make the caller of ParseSenderAddressFromHead responsible for parse errors
There are two callsites for ParseSenderAddressFromHead; one in
RT::Interface::Email::Auth::MailFrom->GetCurrentUser, and one in
RT::Interface::Email->CheckForSuspiciousSender.  The former already
warns if there are no addresses; and in the latter, the lack of
addresses is not a failure.

This reduces the number of warnings generated when there are parse
failures.