Update deployment configurations to remove Alias lines
As /static/ is not a single directory, but an overlayed union of the
plugins, local overrides, and core statici files, a simple Apache
"Alias" line cannot produce the same complexity as the static
middleware.  Remove the Alias lines (and equivalent) and allow RT to
serve the static files itself.

This also has performance implications.  Some simple testing under
Apache with apachebench, with 5000 requests to a single-process server
on localhost with hot caches is below.  Previous to this branch:

                                    | mod_perl | fastcgi
------------------------------------+----------+---------
/NoAuth/RichText/ckeditor.js        |   19.3ms |  17.8ms
/NoAuth/images/farbtastic/wheel.png |   18.3ms |   0.4ms

Only /NoAuth/images/ was served via Alias, and only under fastcgi, due
to a misconfiguration in our suggested deployment -- namely the lack of
'SetHandler default' in a '<Location /NoAuth/images>' section.  The
files were chosen to be the largest that we serve in each category.

After this branch:
                                    | mod_perl | fastcgi
------------------------------------+----------+---------
/static/RichText/ckeditor.js        |    2.9ms |   3.9ms
/static/images/farbtastic/wheel.png |    1.8ms |   1.7ms

That is, this provides an order-of-magnitude speedup on ckeditor.js on
both mod_perl and fastcgi, and an order-of-magnitude speedup under
mod_perl for images, at the cost of an order of magnitude for images
under fastcgi.  However, other clear optimization paths exist for the
static file middleware (Plack::Middleware::XSendfile, cache headers),
which will benefit all deployments.