Ensure that leading newlines in templates are preserved on submit
2d96251 removed a newline after the opening <textarea> -- however, browsers
assume that the first newline in a <textarea> is for formatting, and do not
include it in their submitted value.  This lead to the first newline
 of templates being removed on submit.  See [issues #21152].

The part of the spec that says this is how browsers should behave is
http://www.w3.org/TR/html4/appendix/notes.html#h-B.3.1

The failing tests included with this commit are TODOed. Current versions of
WWW::Mechanize do not act as browsers do in this regard.