Skip CF validation for empty fields which aren't submitted with "Magic"
If a CF value is not submitted and the "Magic" marker is also missing,
skip validation.  The most likely trigger for this is a link containing
only core field query parameters submitted to a page which processes
both core fields and CFs.  By skipping validation in this case, we avoid
nonsensical errors.