Use RT_Config/RT_SiteConfig value for DBA, not configure-time value
This is purely for consistency.  For most cases, DatabaseAdmin in
RT_Config will be what was just installed using 'make upgrade', which
was what was configured via './configure' -- and as such this is no
change.

In the case that the user has configured a different value in
RT_SiteConfig, it will use that value.  While odd, this is at least
consistent with all other configure-time values, which are only pushed
into RT_Config.pm, and read via RT->Config->Get(...) in sbin/ files (not
hard-coded via @FOO@ configure-time substitution).