Show custom field input hints (FriendlyPattern) when editing
Defaults on for usability: it's nicer to note what format to use before
submitting, rather than rejecting after submission.  The hint message is
the same as the validation message at the moment (but could change in
the future), except it renders in black instead of red.  If invalid
input is nevertheless submitted, the error (in red) is shown instead of
the hint.

This makes t/web/cf_pattern.t pass for ticket creation, but the modify
pages still need further fixing.

Adjusts t/web/cf_groupings.t to ensure it matches the result messages
and not the newly added CF hints.