Redirect to the full ticket create page instead of erroring when CFs are required
It's much more friendly to simply redirect users to the page they need
to be on, with their already submitted data prefilled.  This lets them
just fill in the few required CFs and go on their way as quickly as
possible (which is the whole point of quick create anyway!).