Remove the (socket => "inet"), suggestion, based on Sys::Syslog docs
Modern Sys::Syslog stresses that setting setlogsock (which is what
passing "socket" does) is strongly discouraged -- see the first two of
the "THE RULES OF SYS::SYSLOG".  The two reports for Solaris and
UnixWare (tickets #4717 and #5755) were for Sys::Syslog 0.04 or below,
and should no longer be necessary with 0.16 or above.