Push logic into RT::Record->AddRoleMember from _AddWatcher
RT::Ticket and RT::Queue contain very similar code in _AddWatcher, which
attempts to check IsRTAddress and resolve the passed email address to a
user, creating it as necessary -- though RT::Queue's implementation
predates ->LoadOrCreateByEmail.  This should not be surprising, as the
code in RT::Queue was copied wholesale from RT::Ticket in 40b1ea0.

Unify the two divergent codepaths by pushing additional the additional
logic down into RT::Record->AddRoleMember.  This also serves to resolve
a bug when AddRoleMember was called with a non-existant email address.