Reduce duplication by placing ACL callback in AddRoleMember
Both RT::Ticket and RT::Queue's AddWatcher method attempt to resolve the
passed-in email address in order to ascertain if it has appropriate
permissions.  This requires duplicating code in _AddWatcher (now found
in AddRoleMember).  In the case of RT::Queue's AddWatcher code, this
also effectively prevented passing in not-yet-in-the-database email
addresses, as ACLs were checked with ->Load, despite _AddWatcher later
using LoadOrCreateByEmail.

Provide a hook in AddRoleMember after the principal has been resolved,
which can be used to apply ACLs.  This also removes the need for
separation between _AddMember and AddMember.