Remove ShowUserEntry; last callsite removed in RT 3.7
RT 3.7 (in e0c582a) renamed ShowUserEntry to ShowUser, removing all
callsites of it, and (in 0f2a881) replaced ShowUserEntry as a wrapper
around ShowUser.  Meanwhile, RT 3.6 adjusted ShowUserEntry to display
which users were squelched.  When the two branches were merged (in
b5792bd), ShowUserEntry was left as a wrapper around ShowUser which
additionally displayed squelching information -- but was not called from
anywhere.

Remove it, as it is uncalled; ShowUser has evolved since then, and the
places where squelching is relevant now display that information on
their own.