Combine the two "TransactionLoading" configs, Defer and Delay, into one
The new config option is called ShowHistory and may be one of three
values: delay, click (previously defer), and always.  The default is
"delay" since it provides superior page load times.

A single option simplifies user preferences and makes it clear that this
is a "pick one" choice.  You can't have both delayed loading and click
to load.

An upgrade script takes care of porting DeferTransactionLoading
preferences forward.  For preferences which are explicitly false, we
remove the entry and set ShowHistory to "delay".  "always" may
technically be the most correct interpretation, but "delay" is the most
useful and still shows history without any interaction.