Correct relative paths to docs/images/ in POD
Both documents are under docs/customizing/, so the paths were incorrect
to start with.

It's worth noting that image paths in POD are relative to the generated
HTML tree, not the POD document itself.  While docs/ exists at the same
directory level in both source and destination, bin/, sbin/, etc/, and
lib/ are all one level closer to the images directory once turned into
HTML.