Rename RT::Queue->Lifecycle to LifecycleObj for consistency
All other columns which reference other records use ColumnName for the
database value and ColumnNameObj for the inflated object.  Mimic the
same methods in RT::Ticket as well to avoid confusion, despite
RT::Ticket not having a Lifecycle column itself.

This reverts a small part of e904afe which added special handling for
the Lifecycle column to RT::Record->Update.  With the existence of a
LifecycleObj method, the previously existing Name handling works
correctly.

A usage of $QueueObj->Lifecycle in share/html/Admin/Queues/Modify.html
is left as-is to fix a bug: the intention was always to use the
lifecycle name as the current value, but instead an object was passed to
share/html/Widget/Select:InputOnly.  Mason helpfully coerced the object
into an array, which by happenstance led to the correct behaviour.