Refactor Lifecycle methods and behaviours into roles
The first use of roles in RT!  With support for extensions adding
lifecycles on arbitrary object types, it's extremely useful to abstract
the common behaviour patterns into roles that can be consumed by many
classes.

Role::Basic provides a very straightforward implementation without the
prerequisite of shoehorning all our classes into Moose's view of the
world.  The most notable missing feature is parameterized roles.
Instead, roles can require a method in the consuming class which returns
the value of the "parameter".  This does limit individual role
application to one-per-class instead of potentially many instances of
the same parameterized role, but it is a sufficient workaround for now.