RT::Lifecycle->Type accessor
Stashes the "effective" type of the lifecycle at the top level of the
object, including for virtual aggregate lifecycles (i.e. lifecycles
loaded with no name, only a type).  In cases when a lifecycle is
requested by name with a type that conflicts with reality, ->Type will
return the requested type, not the *declared* type of the lifecycle from
the config.  This ensures ->Type matches the blessed lifecycle class;
hence the "effective" type.