Remove misleading ShowTicket rights check in search queue selector
/Elements/SelectQueue only uses CheckQueueRight if ShowAllQueues is
false; by default it is true, however, so the ShowTicket right was never
checked when generating the queue list for the query builder.  This has
been true since the ACL check was introduced in 11195d4.

In fact, checking the ShowTicket right would incorrectly hide queues for
which the current user has ShowTicket via Requestor or another role, but
not globally on the queue.  The ShowTicket right properly belongs at the
individual ticket level, and should not be used for the queue list.