Provide a method to access role attributes rather than using _ROLES directly
This is a bit cleaner, documented, and doesn't let the caller change any
attributes using the returned reference.  The next step would be to
start using objects to represent role metadata, but that feels much
heavier an abstraction than is needed.  A hashref is fine for a
read-only handful of flags.